==> preprocessing_PDB/03. Run PLIP.ipynb <==
# The protein-ligand interaciton profiler (PLIP) is utilized to extract interaction site labels.
# Salentin, S., Schreiber, S., Haupt, V. J., Adasme, M. F., & Schroeder, M. (2015). PLIP: fully automated protein–ligand interaction profiler. Nucleic acids research, 43(W1), W443-W447.

# For services provided via web, it is difficult to apply on large datasets, so we utilize the code provided in the official github below.
# https://github.com/pharmai/plip

# All downloaded compound-protein structures is analyzed by PLIP.
# source files dir: "./data/PDB/protein"

# The results obtain by PLIP is saved in the dir below for further preprocessing to generate input data.
# results files dir: "./data/PLIP_results"

# More detailed PLIP analysis results file examples can be found in the ./data/PLIP_results dir.
