==> preprocessing_PDB/07. Get final data.ipynb <==
import pickle
import numpy as np
import pandas as pd
from rdkit import Chem

### 1. Load data

#### 1.1 Binding affinity data load 

PDBbind_dta_df = pd.read_csv("./preprocessed_data/step5_PDBbind_data.tsv", sep = "\t")
print(f"[PDBbind]: {len(PDBbind_dta_df)}")

CASF2016_dta_df = pd.read_csv("./preprocessed_data/step5_CASF2016_data.tsv", sep = "\t")
print(f"[CASF2016]: {len(CASF2016_dta_df)}")

CASF2013_dta_df = pd.read_csv("./preprocessed_data/step5_CASF2013_data.tsv", sep = "\t")
print(f"[CASF2013]: {len(CASF2013_dta_df)}")

CSAR2014_dta_df = pd.read_csv("./preprocessed_data/step5_CSAR2014_data.tsv", sep = "\t")
print(f"[CSAR2014]: {len(CSAR2014_dta_df)}")

CSAR2012_dta_df = pd.read_csv("./preprocessed_data/step5_CSAR2012_data.tsv", sep = "\t")
print(f"[CSAR2012]: {len(CSAR2012_dta_df)}")

CSARset1_dta_df = pd.read_csv("./preprocessed_data/step5_CSARset1_data.tsv", sep = "\t")
print(f"[CSARset1]: {len(CSARset1_dta_df)}")

CSARset2_dta_df = pd.read_csv("./preprocessed_data/step5_CSARset2_data.tsv", sep = "\t")
print(f"[CSARset2]: {len(CSARset2_dta_df)}")

Astex_dta_df = pd.read_csv("./preprocessed_data/step5_Astex_data.tsv", sep = "\t")
print(f"[Astex]: {len(Astex_dta_df)}")

COACH420_dta_df = pd.read_csv("./preprocessed_data/step5_COACH420_data.tsv", sep = "\t")
print(f"[COACH420]: {len(COACH420_dta_df)}")

HOLO4K_dta_df = pd.read_csv("./preprocessed_data/step5_HOLO4K_data.tsv", sep = "\t")
print(f"[HOLO4K]: {len(HOLO4K_dta_df)}")

#### 1.2 Interacton site data load

with open("./preprocessed_data/step6_PDBbind_data.pkl", "rb") as f:
    PDBbind_interacton_results = pickle.load(f)
print(f"[PDBbind]: {len(PDBbind_interacton_results)}")

with open("./preprocessed_data/step6_CASF2016_data.pkl", "rb") as f:
    CASF2016_interacton_results = pickle.load(f)
print(f"[CASF2016]: {len(CASF2016_interacton_results)}")

with open("./preprocessed_data/step6_CASF2013_data.pkl", "rb") as f:
    CASF2013_interacton_results = pickle.load(f)
print(f"[CASF2013]: {len(CASF2013_interacton_results)}")

with open("./preprocessed_data/step6_CSAR2014_data.pkl", "rb") as f:
    CSAR2014_interacton_results = pickle.load(f)
print(f"[CSAR2014]: {len(CSAR2014_interacton_results)}")

with open("./preprocessed_data/step6_CSAR2012_data.pkl", "rb") as f:
    CSAR2012_interacton_results = pickle.load(f)
print(f"[CSAR2012]: {len(CSAR2012_interacton_results)}")

with open("./preprocessed_data/step6_CSARset1_data.pkl", "rb") as f:
    CSARset1_interacton_results = pickle.load(f)
print(f"[CSARset1]: {len(CSARset1_interacton_results)}")

with open("./preprocessed_data/step6_CSARset2_data.pkl", "rb") as f:
    CSARset2_interacton_results = pickle.load(f)
print(f"[CSARset2]: {len(CSARset2_interacton_results)}")

with open("./preprocessed_data/step6_Astex_data.pkl", "rb") as f:
    Astex_interacton_results = pickle.load(f)
print(f"[Astex]: {len(Astex_interacton_results)}")

with open("./preprocessed_data/step6_COACH420_data.pkl", "rb") as f:
    COACH420_interacton_results = pickle.load(f)
print(f"[COACH420]: {len(COACH420_interacton_results)}")

with open("./preprocessed_data/step6_HOLO4K_data.pkl", "rb") as f:
    HOLO4K_interacton_results = pickle.load(f)
print(f"[HOLO4K]: {len(HOLO4K_interacton_results)}")

#### 1.3 Binding site data load

scPDB_bs_df = pd.read_csv("./preprocessed_bs_data/step4_scPDB_data.tsv", sep = "\t")
print(f"[scPDB]: {len(scPDB_bs_df)}")

PDBbind_bs_df = pd.read_csv("./preprocessed_bs_data/step4_PDBbind_data.tsv", sep = "\t")
print(f"[PDBbind]: {len(PDBbind_bs_df)}")

CASF2016_bs_df = pd.read_csv("./preprocessed_bs_data/step4_CASF2016_data.tsv", sep = "\t")
print(f"[CASF2016]: {len(CASF2016_bs_df)}")

CASF2013_bs_df = pd.read_csv("./preprocessed_bs_data/step4_CASF2013_data.tsv", sep = "\t")
print(f"[CASF2013]: {len(CASF2013_bs_df)}")

CSAR2014_bs_df = pd.read_csv("./preprocessed_bs_data/step4_CSAR2014_data.tsv", sep = "\t")
print(f"[CSAR2014]: {len(CSAR2014_bs_df)}")

CSAR2012_bs_df = pd.read_csv("./preprocessed_bs_data/step4_CSAR2012_data.tsv", sep = "\t")
print(f"[CSAR2012]: {len(CSAR2012_bs_df)}")

CSARset1_bs_df = pd.read_csv("./preprocessed_bs_data/step4_CSARset1_data.tsv", sep = "\t")
print(f"[CSARset1]: {len(CSARset1_bs_df)}")

CSARset2_bs_df = pd.read_csv("./preprocessed_bs_data/step4_CSARset2_data.tsv", sep = "\t")
print(f"[CSARset2]: {len(CSARset2_bs_df)}")

Astex_bs_df = pd.read_csv("./preprocessed_bs_data/step4_Astex_data.tsv", sep = "\t")
print(f"[Astex]: {len(Astex_bs_df)}")

COACH420_bs_df = pd.read_csv("./preprocessed_bs_data/step4_COACH420_data.tsv", sep = "\t")
print(f"[COACH420]: {len(COACH420_bs_df)}")

HOLO4K_bs_df = pd.read_csv("./preprocessed_bs_data/step4_HOLO4K_data.tsv", sep = "\t")
print(f"[HOLO4K]: {len(HOLO4K_bs_df)}")

### 2. Preprocessing results

#### 2.1 DTA data remove interactions over length

def remove_over_lengths(dict_):
    remove_keys_list = list()
    
    for key in list(dict_.keys()):
        seq_len = len(dict_[key]['uniprot_seq'])
        
        if seq_len > 1500:
            remove_keys_list.append(key)
            
    for key in remove_keys_list:
        del dict_[key]
    
    return dict_

PDBbind_interacton_results = remove_over_lengths(PDBbind_interacton_results)
print(len(PDBbind_interacton_results))

CASF2016_interacton_results = remove_over_lengths(CASF2016_interacton_results)
print(len(CASF2016_interacton_results))

CASF2013_interacton_results = remove_over_lengths(CASF2013_interacton_results)
print(len(CASF2013_interacton_results))

CSAR2014_interacton_results = remove_over_lengths(CSAR2014_interacton_results)
print(len(CSAR2014_interacton_results))

CSAR2012_interacton_results = remove_over_lengths(CSAR2012_interacton_results)
print(len(CSAR2012_interacton_results))

CSARset1_interacton_results = remove_over_lengths(CSARset1_interacton_results)
print(len(CSARset1_interacton_results))

CSARset2_interacton_results = remove_over_lengths(CSARset2_interacton_results)
print(len(CSARset2_interacton_results))

Astex_interacton_results = remove_over_lengths(Astex_interacton_results)
print(len(Astex_interacton_results))

COACH420_interacton_results = remove_over_lengths(COACH420_interacton_results)
print(len(COACH420_interacton_results))

HOLO4K_interacton_results = remove_over_lengths(HOLO4K_interacton_results)
print(len(HOLO4K_interacton_results))

#### 2.2 Remove zero PLIP

def check_interaction(input_dict):
    remove_keys_list = list()
    
    for key in list(input_dict.keys()):

        atom_bond = input_dict[key]['atom_bond_type']
        residue_bond = input_dict[key]['residue_bond_type']

        if len(atom_bond) == 0 or len(residue_bond) == 0:
            remove_keys_list.append(key)
    
    for key in remove_keys_list:
        del input_dict[key]
    
    return input_dict

PDBbind_interacton_results = check_interaction(PDBbind_interacton_results)
print(len(PDBbind_interacton_results))

CASF2016_interacton_results = check_interaction(CASF2016_interacton_results)
print(len(CASF2016_interacton_results))

CASF2013_interacton_results = check_interaction(CASF2013_interacton_results)
print(len(CASF2013_interacton_results))

CSAR2014_interacton_results = check_interaction(CSAR2014_interacton_results)
print(len(CSAR2014_interacton_results))

CSAR2012_interacton_results = check_interaction(CSAR2012_interacton_results)
print(len(CSAR2012_interacton_results))

CSARset1_interacton_results = check_interaction(CSARset1_interacton_results)
print(len(CSARset1_interacton_results))

CSARset2_interacton_results = check_interaction(CSARset2_interacton_results)
print(len(CSARset2_interacton_results))

Astex_interacton_results = check_interaction(Astex_interacton_results)
print(len(Astex_interacton_results))

COACH420_interacton_results = check_interaction(COACH420_interacton_results)
print(len(COACH420_interacton_results))

HOLO4K_interacton_results = check_interaction(HOLO4K_interacton_results)
print(len(HOLO4K_interacton_results))

#### 2.3 Binding site data remove protein over length

scPDB_bs_df = scPDB_bs_df[scPDB_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
scPDB_bs_df = scPDB_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(scPDB_bs_df))

PDBbind_bs_df = PDBbind_bs_df[PDBbind_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
PDBbind_bs_df = PDBbind_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(PDBbind_bs_df))

CASF2016_bs_df = CASF2016_bs_df[CASF2016_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
CASF2016_bs_df = CASF2016_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(CASF2016_bs_df))

CASF2013_bs_df = CASF2013_bs_df[CASF2013_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
CASF2013_bs_df = CASF2013_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(CASF2013_bs_df))

CSAR2014_bs_df = CSAR2014_bs_df[CSAR2014_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
CSAR2014_bs_df = CSAR2014_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(CSAR2014_bs_df))

CSAR2012_bs_df = CSAR2012_bs_df[CSAR2012_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
CSAR2012_bs_df = CSAR2012_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(CSAR2012_bs_df))

CSARset1_bs_df = CSARset1_bs_df[CSARset1_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
CSARset1_bs_df = CSARset1_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(CSARset1_bs_df))

CSARset2_bs_df = CSARset2_bs_df[CSARset2_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
CSARset2_bs_df = CSARset2_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(CSARset2_bs_df))

Astex_bs_df = Astex_bs_df[Astex_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
Astex_bs_df = Astex_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(Astex_bs_df))

COACH420_bs_df = COACH420_bs_df[COACH420_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
COACH420_bs_df = COACH420_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(COACH420_bs_df))

HOLO4K_bs_df = HOLO4K_bs_df[HOLO4K_bs_df['Uniprot_Seqs'].apply(len) <= 1500].reset_index(drop=True)
HOLO4K_bs_df = HOLO4K_bs_df.dropna(axis = 0).reset_index(drop=True)
print(len(HOLO4K_bs_df))

#### 2.4 Check DTA and binding sites data

def cross_check(input_dict, input_list):
    remove_keys_list = list()
    
    for key in list(input_dict.keys()):
        if key not in input_list:
            remove_keys_list.append(key)

    for key in remove_keys_list:
        del input_dict[key]
    
    return input_dict

PDBbind_bs_list = [f"{i}_{j}" for i, j in zip(PDBbind_bs_df.PDB_IDs.values, PDBbind_bs_df.Lig_codes.values)]
PDBbind_interacton_results = cross_check(PDBbind_interacton_results, PDBbind_bs_list)
len(PDBbind_interacton_results)

CASF2016_bs_list = [f"{i}_{j}" for i, j in zip(CASF2016_bs_df.PDB_IDs.values, CASF2016_bs_df.Lig_codes.values)]
CASF2016_interacton_results = cross_check(CASF2016_interacton_results, CASF2016_bs_list)
len(CASF2016_interacton_results)

CASF2013_bs_list = [f"{i}_{j}" for i, j in zip(CASF2013_bs_df.PDB_IDs.values, CASF2013_bs_df.Lig_codes.values)]
CASF2013_interacton_results = cross_check(CASF2013_interacton_results, CASF2013_bs_list)
len(CASF2013_interacton_results)

CSAR2014_bs_list = [f"{i}_{j}" for i, j in zip(CSAR2014_bs_df.PDB_IDs.values, CSAR2014_bs_df.Lig_codes.values)]
CSAR2014_interacton_results = cross_check(CSAR2014_interacton_results, CSAR2014_bs_list)
len(CSAR2014_interacton_results)

CSAR2012_bs_list = [f"{i}_{j}" for i, j in zip(CSAR2012_bs_df.PDB_IDs.values, CSAR2012_bs_df.Lig_codes.values)]
CSAR2012_interacton_results = cross_check(CSAR2012_interacton_results, CSAR2012_bs_list)
len(CSAR2012_interacton_results)

CSARset1_bs_list = [f"{i}_{j}" for i, j in zip(CSARset1_bs_df.PDB_IDs.values, CSARset1_bs_df.Lig_codes.values)]
CSARset1_interacton_results = cross_check(CSARset1_interacton_results, CSARset1_bs_list)
len(CSARset1_interacton_results)

CSARset2_bs_list = [f"{i}_{j}" for i, j in zip(CSARset2_bs_df.PDB_IDs.values, CSARset2_bs_df.Lig_codes.values)]
CSARset2_interacton_results = cross_check(CSARset2_interacton_results, CSARset2_bs_list)
len(CSARset2_interacton_results)

Astex_bs_list = [f"{i}_{j}" for i, j in zip(Astex_bs_df.PDB_IDs.values, Astex_bs_df.Lig_codes.values)]
Astex_interacton_results = cross_check(Astex_interacton_results, Astex_bs_list)
len(Astex_interacton_results)

COACH420_bs_list = [f"{i}_{j}" for i, j in zip(COACH420_bs_df.PDB_IDs.values, COACH420_bs_df.Lig_codes.values)]
COACH420_interacton_results = cross_check(COACH420_interacton_results, COACH420_bs_list)
len(COACH420_interacton_results)

HOLO4K_bs_list = [f"{i}_{j}" for i, j in zip(HOLO4K_bs_df.PDB_IDs.values, HOLO4K_bs_df.Lig_codes.values)]
HOLO4K_interacton_results = cross_check(HOLO4K_interacton_results, HOLO4K_bs_list)
len(HOLO4K_interacton_results)

### 3. Remove duplicate interaction in binding site data

def get_list(df):
    results = list()
    
    for line in df.values:
        pdbid, lig_code = line[0], line[1]
        results.append(f"{pdbid}_{lig_code}")
    return results

scPDB_bs_list = get_list(scPDB_bs_df)
print(len(scPDB_bs_list))

PDBbind_bs_list = get_list(PDBbind_bs_df)
print(len(PDBbind_bs_list))

CASF2016_bs_list = get_list(CASF2016_bs_df)
print(len(CASF2016_bs_list))

CASF2013_bs_list = get_list(CASF2013_bs_df)
print(len(CASF2013_bs_list))

CSAR2014_bs_list = get_list(CSAR2014_bs_df)
print(len(CSAR2014_bs_list))

CSAR2012_bs_list = get_list(CSAR2012_bs_df)
print(len(CSAR2012_bs_list))

CSARset1_bs_list = get_list(CSARset1_bs_df)
print(len(CSARset1_bs_list))

CSARset2_bs_list = get_list(CSARset2_bs_df)
print(len(CSARset2_bs_list))

Astex_bs_list = get_list(Astex_bs_df)
print(len(Astex_bs_list))

COACH420_bs_list = get_list(COACH420_bs_df)
print(len(COACH420_bs_list))

HOLO4K_bs_list = get_list(HOLO4K_bs_df)
print(len(HOLO4K_bs_list))

BS_training = list(set(scPDB_bs_list) | set(PDBbind_bs_list))
print(len(BS_training))

BS_training = list(set(BS_training) - set(CASF2016_bs_list) - set(CASF2013_bs_list) - set(CSAR2014_bs_list) 
                   - set(CSAR2012_bs_list) - set(CSARset1_bs_list) - set(CSARset2_bs_list) 
                    - set(Astex_bs_list) - set(COACH420_bs_list) - set(HOLO4K_bs_list))
print(len(BS_training))

### 4. Make DTA dataset

def make_dta_dataset(df, interaction_results):
    results = {"PDB_IDs":[], "Uniprot_IDs":[], "Ligand_Codes":[], "Labels":[], "Uniprot_Seqs":[], "ISO_SMILES":[]}
    pdbid_list, lig_code_list, label_list, smiles_iso_list = df.PDB.values, df.Lig_code.values, df.Labels.values, df.RDKit_iso_SMILES.values
    
    for pdbid, lig_code, label, iso_smiles in zip(pdbid_list, lig_code_list, label_list, smiles_iso_list):
        
        if f"{pdbid}_{lig_code}" in interaction_results:
        
            uniprot_id, uniprot_seq = interaction_results[f"{pdbid}_{lig_code}"]['uniprot_id'], interaction_results[f"{pdbid}_{lig_code}"]['uniprot_seq']
            
            results["PDB_IDs"].append(pdbid)
            results['Uniprot_IDs'].append(uniprot_id)
            results['Ligand_Codes'].append(lig_code)
            results['Labels'].append(label)
            results['Uniprot_Seqs'].append(uniprot_seq)
            results['ISO_SMILES'].append(iso_smiles)
    
    return pd.DataFrame(results)

PDBbind_dta_df = make_dta_dataset(PDBbind_dta_df, PDBbind_interacton_results)
print(len(PDBbind_dta_df))

CASF2016_dta_df = make_dta_dataset(CASF2016_dta_df, CASF2016_interacton_results)
print(len(CASF2016_dta_df))

CASF2013_dta_df = make_dta_dataset(CASF2013_dta_df, CASF2013_interacton_results)
print(len(CASF2013_dta_df))

CSAR2014_dta_df = make_dta_dataset(CSAR2014_dta_df, CSAR2014_interacton_results)
print(len(CSAR2014_dta_df))

CSAR2012_dta_df = make_dta_dataset(CSAR2012_dta_df, CSAR2012_interacton_results)
print(len(CSAR2012_dta_df))

CSARset1_dta_df = make_dta_dataset(CSARset1_dta_df, CSARset1_interacton_results)
print(len(CSARset1_dta_df))

CSARset2_dta_df = make_dta_dataset(CSARset2_dta_df, CSARset2_interacton_results)
print(len(CSARset2_dta_df))

Astex_dta_df = make_dta_dataset(Astex_dta_df, Astex_interacton_results)
print(len(Astex_dta_df))

def make_dta_dataset(df, interaction_results):
    results = {"PDB_IDs":[], "Uniprot_IDs":[], "Ligand_Codes":[], "Uniprot_Seqs":[], "ISO_SMILES":[]}
    pdbid_list, lig_code_list, smiles_iso_list = df.PDB.values, df.Lig_code.values, df.RDKit_iso_SMILES.values
    
    for pdbid, lig_code, iso_smiles in zip(pdbid_list, lig_code_list, smiles_iso_list):
        
        if f"{pdbid}_{lig_code}" in interaction_results:
        
            uniprot_id, uniprot_seq = interaction_results[f"{pdbid}_{lig_code}"]['uniprot_id'], interaction_results[f"{pdbid}_{lig_code}"]['uniprot_seq']
            
            results["PDB_IDs"].append(pdbid)
            results['Uniprot_IDs'].append(uniprot_id)
            results['Ligand_Codes'].append(lig_code)
            results['Uniprot_Seqs'].append(uniprot_seq)
            results['ISO_SMILES'].append(iso_smiles)
    
    return pd.DataFrame(results)

COACH420_IS_df = make_dta_dataset(COACH420_dta_df, COACH420_interacton_results)
print(len(COACH420_IS_df))

HOLO4K_IS_df = make_dta_dataset(HOLO4K_dta_df, HOLO4K_interacton_results)
print(len(HOLO4K_IS_df))

PDBbind_dta_df.to_csv("../input_data/PDB/BA/Training_BA_data.tsv", sep = "\t", index = False)
CASF2016_dta_df.to_csv("../input_data/PDB/BA/CASF2016_BA_data.tsv", sep = "\t", index = False)
CASF2013_dta_df.to_csv("../input_data/PDB/BA/CASF2013_BA_data.tsv", sep = "\t", index = False)
CSAR2014_dta_df.to_csv("../input_data/PDB/BA/CSAR2014_BA_data.tsv", sep = "\t", index = False)
CSAR2012_dta_df.to_csv("../input_data/PDB/BA/CSAR2012_BA_data.tsv", sep = "\t", index = False)
CSARset1_dta_df.to_csv("../input_data/PDB/BA/CSARset1_BA_data.tsv", sep = "\t", index = False)
CSARset2_dta_df.to_csv("../input_data/PDB/BA/CSARset2_BA_data.tsv", sep = "\t", index = False)
Astex_dta_df.to_csv("../input_data/PDB/BA/Astex_BA_data.tsv", sep = "\t", index = False)
COACH420_IS_df.to_csv("../input_data/PDB/BA/COACH420_IS_data.tsv", sep = "\t", index = False)
HOLO4K_IS_df.to_csv("../input_data/PDB/BA/HOLO4K_IS_data.tsv", sep = "\t", index = False)

### 5. Make Interaction dataset

def get_pairwise_label_nonh(pdbid, pid, cid, interaction_dict, mol):
    if pdbid in interaction_dict:
        sdf_element = np.array([atom.GetSymbol().upper() for atom in mol.GetAtoms()])
        atom_element = np.array(interaction_dict[pdbid]['atom_element'], dtype=str)
        atom_name_list = np.array(interaction_dict[pdbid]['atom_name'], dtype=str)
        atom_interact = np.array(interaction_dict[pdbid]['atom_interact'], dtype=int)
        nonH_position = np.where(atom_element != ('H'))[0]
        
        assert sum(atom_element[nonH_position] != sdf_element) == 0
        
        atom_name_list = atom_name_list[nonH_position].tolist()

        pairwise_mat = np.zeros((len(nonH_position), len(interaction_dict[pdbid]['uniprot_seq'])), dtype=np.int32)
        
        for atom_name, bond_type in interaction_dict[pdbid]['atom_bond_type']:
            atom_idx = atom_name_list.index(str(atom_name))
            assert atom_idx < len(nonH_position)

            for seq_idx, bond_type_seq in interaction_dict[pdbid]['residue_bond_type']:
                if bond_type == bond_type_seq:
                    pairwise_mat[atom_idx, seq_idx] = 1
                    
        if len(np.where(pairwise_mat != 0)[0]) != 0:
            pairwise_mask = True
            return True, pairwise_mat
        
    return False, np.zeros((1,1))

def make_pairwie_labels(interaction_sites_dict):
    
    results_dict, convert_dict = dict(), dict()
    
    path = "./data/PDB/ligand"

    for idx, key in enumerate(list(interaction_sites_dict.keys())):

        pdbid, cid, uniprotid = key.split("_")[0], key.split("_")[1], interaction_sites_dict[key]['uniprot_id']

        sdf_path = f"{path}/{cid}_ideal.sdf"

        suppl = Chem.SDMolSupplier(sdf_path)
        mol = next(iter(suppl))

        inter_data = get_pairwise_label_nonh(key, uniprotid, cid, interaction_sites_dict, mol)

        if f"{uniprotid}_{cid}" in results_dict:
            prev_inter_data = np.sum(results_dict[f"{uniprotid}_{cid}"])

            if prev_inter_data < np.sum(inter_data[1]):
                results_dict[f"{uniprotid}_{cid}"] = inter_data[1]

        else:
            results_dict[f"{uniprotid}_{cid}"] = inter_data[1]

        convert_dict[key] = f"{uniprotid}_{cid}"
    
    print(len(results_dict), len(convert_dict))

    final_dict = dict()
    for key in list(interaction_sites_dict.keys()):
        final_dict[key] = results_dict[convert_dict[key]]
        
    return final_dict

PDBbind_nonh_interactions_dict = make_pairwie_labels(PDBbind_interacton_results)
print(len(PDBbind_nonh_interactions_dict))

def make_pairwie_labels(interaction_sites_dict):
    
    results_dict, convert_dict = dict(), dict()
    
    path = "./data/PDB/ligand"

    for key in list(interaction_sites_dict.keys()):

        pdbid, cid, uniprotid = key.split("_")[0], key.split("_")[1], interaction_sites_dict[key]['uniprot_id']
        sdf_path = f"{path}/{cid}_ideal.sdf"
        
        suppl = Chem.SDMolSupplier(sdf_path)
        mol = next(iter(suppl))

        results_dict[f"{pdbid}_{cid}"] = get_pairwise_label_nonh(key, uniprotid, cid, interaction_sites_dict, mol)[1]
    
    return results_dict

CASF2016_nonh_interactions_dict = make_pairwie_labels(CASF2016_interacton_results)
print(len(CASF2016_nonh_interactions_dict))

CASF2013_nonh_interactions_dict = make_pairwie_labels(CASF2013_interacton_results)
print(len(CASF2013_nonh_interactions_dict))

CSAR2014_nonh_interactions_dict = make_pairwie_labels(CSAR2014_interacton_results)
print(len(CSAR2014_nonh_interactions_dict))

CSAR2012_nonh_interactions_dict = make_pairwie_labels(CSAR2012_interacton_results)
print(len(CSAR2012_nonh_interactions_dict))

CSARset1_nonh_interactions_dict = make_pairwie_labels(CSARset1_interacton_results)
print(len(CSARset1_nonh_interactions_dict))

CSARset2_nonh_interactions_dict = make_pairwie_labels(CSARset2_interacton_results)
print(len(CSARset2_nonh_interactions_dict))

Astex_nonh_interactions_dict = make_pairwie_labels(Astex_interacton_results)
print(len(Astex_nonh_interactions_dict))

COACH420_nonh_interactions_dict = make_pairwie_labels(COACH420_interacton_results)
print(len(COACH420_nonh_interactions_dict))

HOLO4K_nonh_interactions_dict = make_pairwie_labels(HOLO4K_interacton_results)
print(len(HOLO4K_nonh_interactions_dict))

with open("../input_data/PDB/Interaction_sites/Training_nonh_interactions.pkl", "wb") as f:
    pickle.dump(PDBbind_nonh_interactions_dict, f)
    
with open("../input_data/PDB/Interaction_sites/CASF2016_nonh_interactions.pkl", "wb") as f:
    pickle.dump(CASF2016_nonh_interactions_dict, f)
    
with open("../input_data/PDB/Interaction_sites/CASF2013_nonh_interactions.pkl", "wb") as f:
    pickle.dump(CASF2013_nonh_interactions_dict, f)
    
with open("../input_data/PDB/Interaction_sites/CSAR2014_nonh_interactions.pkl", "wb") as f:
    pickle.dump(CSAR2014_nonh_interactions_dict, f)
    
with open("../input_data/PDB/Interaction_sites/CSAR2012_nonh_interactions.pkl", "wb") as f:
    pickle.dump(CSAR2012_nonh_interactions_dict, f)
    
with open("../input_data/PDB/Interaction_sites/CSARset1_nonh_interactions.pkl", "wb") as f:
    pickle.dump(CSARset1_nonh_interactions_dict, f)
    
with open("../input_data/PDB/Interaction_sites/CSARset2_nonh_interactions.pkl", "wb") as f:
    pickle.dump(CSARset2_nonh_interactions_dict, f)
    
with open("../input_data/PDB/Interaction_sites/Astex_nonh_interactions.pkl", "wb") as f:
    pickle.dump(Astex_nonh_interactions_dict, f)
    
with open("../input_data/PDB/Interaction_sites/COACH420_nonh_interactions.pkl", "wb") as f:
    pickle.dump(COACH420_nonh_interactions_dict, f)
    
with open("../input_data/PDB/Interaction_sites/HOLO4K_nonh_interactions.pkl", "wb") as f:
    pickle.dump(HOLO4K_nonh_interactions_dict, f)

### 6. Make Binding sites dataset

def convert_bs(input_list):
    try:
        tmp = ",".join(input_list)
        tmp = set(tmp.split(","))

        tmp = sorted(list(map(int, tmp)))
        tmp = list(map(str, tmp))
    except:
        return None
    
    return ",".join(tmp)
    
def make_bs_df(scPDB_df, PDBbind_df, training_list):
    results = {"PDB_IDs":[], "Uniprot_IDs":[], "Uniprot_Seqs":[], "BS_8A":[], "BS_4A":[], "Ligand_Codes":[]}

    tmp = dict()
    for line in scPDB_df.values:
        pdbid, lig_code, uniprot_id, uniprot_seqs, bs_8A, bs_4A = line[0], line[1], line[3], line[4], line[5], line[6]
        
        if len(uniprot_seqs) <= 1500:
        
            if f"{pdbid}_{lig_code}" in training_list:
                if f"{pdbid}_{lig_code}" in tmp:
                    tmp[f"{pdbid}_{lig_code}"]["lig_code"].append(lig_code)
                    tmp[f"{pdbid}_{lig_code}"]["bs_8A"].append(bs_8A)
                    tmp[f"{pdbid}_{lig_code}"]["bs_4A"].append(bs_4A) 
                    
                else:
                    tmp[f"{pdbid}_{lig_code}"] = dict()
                    tmp[f"{pdbid}_{lig_code}"]["uniprot_id"] = uniprot_id
                    tmp[f"{pdbid}_{lig_code}"]["uniprot_seqs"] = uniprot_seqs
                    tmp[f"{pdbid}_{lig_code}"]["lig_code"] = [lig_code]
                    tmp[f"{pdbid}_{lig_code}"]["bs_8A"] = [bs_8A]
                    tmp[f"{pdbid}_{lig_code}"]["bs_4A"] = [bs_4A]
                    
                
    for line in PDBbind_df.values:
        pdbid, lig_code, uniprot_id, uniprot_seqs, bs_8A, bs_4A = line[0], line[1], line[3], line[4], line[5], line[6]
        if len(uniprot_seqs) <= 1500:
            if f"{pdbid}_{lig_code}" in training_list:
                if f"{pdbid}_{lig_code}" in tmp: 
                    tmp[f"{pdbid}_{lig_code}"]["lig_code"].append(lig_code)
                    tmp[f"{pdbid}_{lig_code}"]["bs_8A"].append(bs_8A)
                    tmp[f"{pdbid}_{lig_code}"]["bs_4A"].append(bs_4A) 
                    
                else:
                    tmp[f"{pdbid}_{lig_code}"] = dict()
                    tmp[f"{pdbid}_{lig_code}"]["uniprot_id"] = uniprot_id
                    tmp[f"{pdbid}_{lig_code}"]["uniprot_seqs"] = uniprot_seqs
                    tmp[f"{pdbid}_{lig_code}"]["lig_code"] = [lig_code]
                    tmp[f"{pdbid}_{lig_code}"]["bs_8A"] = [bs_8A]
                    tmp[f"{pdbid}_{lig_code}"]["bs_4A"] = [bs_4A] 
    
    for key in list(tmp.keys()):
        
        lig_code = ",".join(tmp[key]["lig_code"])

        bs_8A, bs_4A = convert_bs(tmp[key]["bs_8A"]), convert_bs(tmp[key]["bs_4A"])
        
        results["PDB_IDs"].append(key.split("_")[0])
        results["Uniprot_IDs"].append(tmp[key]["uniprot_id"])
        results["Uniprot_Seqs"].append(tmp[key]["uniprot_seqs"])
        
        results["BS_8A"].append(bs_8A)
        results["BS_4A"].append(bs_4A)
        results["Ligand_Codes"].append(lig_code)
    
    return pd.DataFrame(results)

BS_training_df = make_bs_df(scPDB_bs_df, PDBbind_bs_df, BS_training)
BS_training_df

def extract_df(df):
    results = {"PDB_IDs":[], "Uniprot_IDs":[], "Uniprot_Seqs":[], "BS_8A":[], "BS_4A":[], "Ligand_Codes":[]}
    
    for line in df.values:
        pdbid, lig_code, uniprot_id, uniprot_seqs, bs_8A, bs_4A = line[0], line[1], line[3], line[4], line[5], line[6]
        
        results["PDB_IDs"].append(pdbid)
        results["Uniprot_IDs"].append(uniprot_id)
        results["Uniprot_Seqs"].append(uniprot_seqs)
        
        results["BS_8A"].append(bs_8A)
        results["BS_4A"].append(bs_4A)
        results["Ligand_Codes"].append(lig_code)
    
    return pd.DataFrame(results)

CASF2016_bs_df = extract_df(CASF2016_bs_df)
CASF2013_bs_df = extract_df(CASF2013_bs_df)
CSAR2014_bs_df = extract_df(CSAR2014_bs_df)
CSAR2012_bs_df = extract_df(CSAR2012_bs_df)
CSARset1_bs_df = extract_df(CSARset1_bs_df)
CSARset2_bs_df = extract_df(CSARset2_bs_df)
Astex_bs_df = extract_df(Astex_bs_df)
COACH420_bs_df = extract_df(COACH420_bs_df)
HOLO4K_bs_df = extract_df(HOLO4K_bs_df)

BS_training_df.to_csv("../input_data/PDB/BS/Training_BS_data.tsv", sep = "\t", index = False)
CASF2016_bs_df.to_csv("../input_data/PDB/BS/CASF2016_BS_data.tsv", sep = "\t", index = False)
CASF2013_bs_df.to_csv("../input_data/PDB/BS/CASF2013_BS_data.tsv", sep = "\t", index = False)
CSAR2014_bs_df.to_csv("../input_data/PDB/BS/CSAR2014_BS_data.tsv", sep = "\t", index = False)
CSAR2012_bs_df.to_csv("../input_data/PDB/BS/CSAR2012_BS_data.tsv", sep = "\t", index = False)
CSARset1_bs_df.to_csv("../input_data/PDB/BS/CSARset1_BS_data.tsv", sep = "\t", index = False)
CSARset2_bs_df.to_csv("../input_data/PDB/BS/CSARset2_BS_data.tsv", sep = "\t", index = False)
Astex_bs_df.to_csv("../input_data/PDB/BS/Astex_BS_data.tsv", sep = "\t", index = False)
COACH420_bs_df.to_csv("../input_data/PDB/BS/COACH420_BS_data.tsv", sep = "\t", index = False)
HOLO4K_bs_df.to_csv("../input_data/PDB/BS/HOLO4K_BS_data.tsv", sep = "\t", index = False)
